Guys I changed some things and I'm not done dw
It looks properly bad right now, but the goal is to try have the links that were initially in the sidebar of the home page now in the middle in eight boxes so that it fills up the page a bit more. It's going poorly but progress has been made. Y'all can mess with it if you want.